+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Set RPDOMap values by byte string.
 
+### Changed
+- The signature of the load_firmware method of EthercatNetwork is changed to add the boot_in_app argument.
+
 ### Fixed
 - Issue when connecting to the virtual drive using an EVE CANopen dictionary.