+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - The signature of the load_firmware method of EthercatNetwork is changed to add the boot_in_app argument.
+- The polling sampling precision of the Poller is improved.
 
 ### Fixed
 - Issue when connecting to the virtual drive using an EVE CANopen dictionary.