+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [7.3.2] - 2024-06-05
 
 ### Added
 - Set RPDOMap values by byte string.